modularized logger construction
