+++ server/lib/logger.js
@@ -15,100 +15,6 @@ var fs = require('fs'),
     os = require('os');
 
 /* Writers */
-function HtmlWriter(io, fields) {
-  'use strict';
-
-  this.writeHeader = function writeHeader() {
-    // document head
-    io.write('<html><head>' + os.EOL);
-    io.write('<link rel="stylesheet" href="http://cdnjs.cloudflare.com/ajax/libs/normalize/3.0.1/normalize.min.css">' + os.EOL);
-
-    // styling
-    io.write('<style>');
-    io.write('td {padding:5px;} .rows td:nth-child(3){font-style:italic;} .rows td:nth-child(5){text-transform:uppercase;} .rows td:first-child{font-weight:bold;}');
-    io.write('</style>');
-
-    // document body
-    io.write('</head><body>' + os.EOL);
-    io.write('<table><thead>' + os.EOL);
-    io.write('<tr><th class="columns">' + fields.join('</th><th>') + '</th></tr>' + os.EOL);
-    io.write('</thead><tbody>' + os.EOL);
-  };
-
-  this.writeContent = function writeContent() {
-    var data = Array.prototype.slice.call(arguments)
-      .join('</td><td>');
-    io.write('<tr class="rows"><td>' + data + '</td></tr>' + os.EOL);
-  };
-
-  this.writeFooter = function writeFooter() {
-    io.write('</tbody></table></body></html>' + os.EOL);
-  };
-}
-
-function CsvWriter(io, fields) {
-  'use strict';
-
-  this.writeHeader = function writeHeader() {
-    io.write(fields.join(','));
-  };
-
-  this.writeContent = function writeContent() {
-    var data = Array.prototype.slice.call(arguments)
-      .join(',')
-      .concat(os.EOL);
-    io.write(data);
-  };
-
-  this.writeFooter = function writeFooter() {};
-}
-
-function TabWriter(io, fields) {
-  'use strict';
-
-  this.writeHeader = function writeHeader() {
-    io.write(fields.join('\t'));
-  };
-
-  this.writeContent = function writeContent() {
-    var data = Array.prototype.slice.call(arguments)
-      .join('\t')
-      .concat(os.EOL);
-    io.write(data);
-  };
-
-  this.writeFooter = function writeFooter() {};
-}
-
-function MarkdownWriter(io, fields) {
-  'use strict';
-
-  this.writeHeader = function writeHeader() {
-    var content = '| ' + fields.join(' | ') + ' |' + os.EOL;
-    var decoration = new Array(content.length)
-      .join('-')
-      .concat(os.EOL);
-
-    io.write(decoration);
-    io.write(content);
-    io.write(decoration);
-  };
-
-  this.writeContent = function writeContent() {
-    var data = Array.prototype.slice.call(arguments)
-      .join(' | ');
-    io.write('| ' + data + ' |' + os.EOL);
-  };
-
-  this.writeFooter = function writeFooter() {
-    var content = '| ' + fields.join(' | ') + ' |';
-    var decoration = new Array(content.length)
-      .join('-')
-      .concat(os.EOL);
-    io.write(decoration);
-  };
-}
-
 function getTime() {
   return new Date().toLocaleTimeString();
 }
@@ -121,12 +27,13 @@ module.exports = function Logger (cfg, uuid) {
     throw new Error('No configuration file found!');
   }
 
+  /* import loggers */
   types = {
-    'csv'      : CsvWriter,
-    'html'     : HtmlWriter,
-    'markdown' : MarkdownWriter,
-    'tsv'      : TabWriter,
-    'tab'      : TabWriter,
+    'csv'      : require('./loggers/csv'),
+    'html'     : require('./loggers/html'),
+    'markdown' : require('./loggers/markdown'),
+    'tsv'      : require('./loggers/tab'),
+    'tab'      : require('./loggers/tab'),
   };
 
   headers = [
